backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
beanie>=1.23.0
motor>=3.3.0
pymongo>=4.6.0
python-dotenv>=1.0.0
pydantic[email]>=2.8.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6

